Add STREAM Codecs (STREAM-PR2) (#252)
* * Bump SNAPSHOT versions to 0.18.0-SNAPSHOT
* Add minor updates to ILP-core to support STREAM
* Reorg poms for STREAM
* Add stream-core module


* Moving these files to a future PR


* Remove unused class


* Add STREAM Codecs

* Rename `codec-framework` to `codecs-framework`
* Add `codecs-stream` to the project.


* Add assertj


* Move to assertJ


* Remove stale TODO


* Replace copy/base bytes.

+++ stream-parent/stream-core/src/main/java/org/interledger/stream/StreamPacket.java
@@ -62,7 +62,7 @@ import java.util.List;
 @Immutable
 public interface StreamPacket {
 
-  int VERSION_1 = 1;
+  short VERSION_1 = 1;
 
   /**
    * Get the default builder.
